rename State related to Table
