+++ rusty_lr_parser/src/parser/parser_expanded.rs
@@ -1,26 +1,25 @@
-
 // This file was generated by rustylr 1.31.0
 // This generated code is targeting rusty_lr version 4.0.0.
 // There might be a build error if the version is not matched.
 //
 // ================================User Codes Begin================================
-use crate::parser::args::PatternArgs;
 use crate::parser::args::GrammarArgs;
-use crate::parser::args::RuleDefArgs;
-use crate::parser::args::RuleLineArgs;
 use crate::parser::args::IdentOrLiteral;
+use crate::parser::args::PatternArgs;
 use crate::parser::args::PrecDPrecArgs;
 use crate::parser::args::RecoveredError;
+use crate::parser::args::RuleDefArgs;
+use crate::parser::args::RuleLineArgs;
 use crate::parser::lexer::Lexed;
-use crate::parser::location::Location;
 use crate::parser::location::Located;
+use crate::parser::location::Location;
 use crate::terminalset::TerminalSet;
 use crate::terminalset::TerminalSetItem;
 use proc_macro2::Group;
 use proc_macro2::TokenStream;
 use quote::ToTokens;
-use std::boxed::Box;
 use rusty_lr_core::rule::ReduceType;
+use std::boxed::Box;
 
 // =================================User Codes End=================================
 /*
@@ -31,7 +30,7 @@ use rusty_lr_core::rule::ReduceType;
 
 0: Rule -> ident RuleType colon RuleLines semicolon
 1: RuleType -> parengroup
-2: RuleType -> 
+2: RuleType ->
 3: RuleLines -> RuleLines pipe RuleLine
 4: RuleLines -> RuleLine
 5: RuleLine -> TokenMapped* PrecDef* Action
@@ -72,7 +71,7 @@ use rusty_lr_core::rule::ReduceType;
 40: Pattern -> dollar ident lparen Pattern comma Pattern error rparen
 41: Pattern -> dollar ident lparen Pattern comma Pattern comma error rparen
 42: Action -> bracegroup
-43: Action -> 
+43: Action ->
 44: IdentOrLiteral -> ident
 45: IdentOrLiteral -> byte_literal
 46: IdentOrLiteral -> char_literal
@@ -110,25 +109,25 @@ use rusty_lr_core::rule::ReduceType;
 78: TokenMapped+ -> TokenMapped
 79: TokenMapped+ -> TokenMapped+ TokenMapped
 80: TokenMapped* -> TokenMapped+
-81: TokenMapped* -> 
+81: TokenMapped* ->
 82: PrecDef+ -> PrecDef
 83: PrecDef+ -> PrecDef+ PrecDef
 84: PrecDef* -> PrecDef+
-85: PrecDef* -> 
+85: PrecDef* ->
 86: caret? -> caret
-87: caret? -> 
+87: caret? ->
 88: TerminalSetItem+ -> TerminalSetItem
 89: TerminalSetItem+ -> TerminalSetItem+ TerminalSetItem
 90: TerminalSetItem* -> TerminalSetItem+
-91: TerminalSetItem* -> 
+91: TerminalSetItem* ->
 92: Pattern+ -> Pattern
 93: Pattern+ -> Pattern+ Pattern
 94: Pattern* -> Pattern+
-95: Pattern* -> 
+95: Pattern* ->
 96: $sep(Pattern*, pipe, +) -> Pattern*
 97: $sep(Pattern*, pipe, +) -> $sep(Pattern*, pipe, +) pipe Pattern*
 98: comma? -> comma
-99: comma? -> 
+99: comma? ->
 100: [^semicolon] -> ident
 101: [^semicolon] -> colon
 102: [^semicolon] -> pipe
@@ -182,20 +181,15 @@ use rusty_lr_core::rule::ReduceType;
 // =============================Generated Codes Begin==============================
 /// type alias for `Context`
 #[allow(non_camel_case_types, dead_code)]
-pub type GrammarContext = ::rusty_lr_core::parser::deterministic::Context<
-    GrammarParser,
-    GrammarDataStack,
-    u8,
->;
+pub type GrammarContext =
+    ::rusty_lr_core::parser::deterministic::Context<GrammarParser, GrammarDataStack, u8>;
 /// type alias for CFG production rule
 #[allow(non_camel_case_types, dead_code)]
-pub type GrammarRule = ::rusty_lr_core::rule::ProductionRule<
-    GrammarTerminalClasses,
-    GrammarNonTerminals,
->;
+pub type GrammarRule =
+    ::rusty_lr_core::rule::ProductionRule<GrammarTerminalClasses, GrammarNonTerminals>;
 /// type alias for runtime parser tables
 #[allow(non_camel_case_types, dead_code)]
-pub type GrammarTables = ::rusty_lr_core::parser::state::DenseFlatTables<
+pub type GrammarTables = ::rusty_lr_core::parser::table::SparseFlatTables<
     GrammarTerminalClasses,
     GrammarNonTerminals,
     u8,
@@ -217,7 +211,7 @@ pub type GrammarParseError = ::rusty_lr_core::parser::deterministic::ParseError<
     std::cmp::PartialEq,
     std::cmp::Eq,
     std::cmp::PartialOrd,
-    std::cmp::Ord
+    std::cmp::Ord,
 )]
 #[repr(usize)]
 pub enum GrammarTerminalClasses {
@@ -270,7 +264,9 @@ impl GrammarTerminalClasses {
     #[inline]
     pub fn from_usize(value: usize) -> Self {
         debug_assert!(
-            value < 44usize, "Terminal class index {} is out of bounds (max {})", value,
+            value < 44usize,
+            "Terminal class index {} is out of bounds (max {})",
+            value,
             44usize
         );
         unsafe { ::std::mem::transmute(value) }
@@ -337,9 +333,7 @@ impl ::rusty_lr_core::parser::terminalclass::TerminalClass for GrammarTerminalCl
             GrammarTerminalClasses::plus
             | GrammarTerminalClasses::star
             | GrammarTerminalClasses::question
-            | GrammarTerminalClasses::exclamation => {
-                ::rusty_lr_core::parser::Precedence::new(1)
-            }
+            | GrammarTerminalClasses::exclamation => ::rusty_lr_core::parser::Precedence::new(1),
             GrammarTerminalClasses::eof => {
                 unreachable!("eof token cannot be used in precedence levels")
             }
@@ -415,7 +409,7 @@ impl std::fmt::Debug for GrammarTerminalClasses {
     std::cmp::PartialEq,
     std::cmp::Eq,
     std::cmp::PartialOrd,
-    std::cmp::Ord
+    std::cmp::Ord,
 )]
 #[repr(usize)]
 pub enum GrammarNonTerminals {
@@ -454,7 +448,9 @@ impl GrammarNonTerminals {
     #[inline]
     pub fn from_usize(value: usize) -> Self {
         debug_assert!(
-            value < 30usize, "Non-terminal index {} is out of bounds (max {})", value,
+            value < 30usize,
+            "Non-terminal index {} is out of bounds (max {})",
+            value,
             30usize
         );
         unsafe { ::std::mem::transmute(value) }
@@ -507,9 +503,7 @@ impl ::rusty_lr_core::parser::nonterminal::NonTerminal for GrammarNonTerminals {
             GrammarNonTerminals::Augmented => "Augmented",
         }
     }
-    fn nonterm_type(
-        &self,
-    ) -> Option<::rusty_lr_core::parser::nonterminal::NonTerminalType> {
+    fn nonterm_type(&self) -> Option<::rusty_lr_core::parser::nonterminal::NonTerminalType> {
         match self {
             GrammarNonTerminals::Rule => None,
             GrammarNonTerminals::RuleType => None,
@@ -612,7 +606,9 @@ pub struct GrammarDataStack {
 }
 impl Default for GrammarDataStack {
     fn default() -> Self {
-        Self { __stack: Vec::new() }
+        Self {
+            __stack: Vec::new(),
+        }
     }
 }
 #[rustfmt::skip]
@@ -6325,7 +6321,13 @@ impl ::rusty_lr_core::parser::data_stack::DataStack for GrammarDataStack {
 ///
 /// Since this struct only holds `'static` references to shared, read-only static parser tables,
 /// it is extremely cheap to instantiate, copy, or clone, and takes very little space.
-#[allow(unused_braces, unused_parens, unused_variables, non_snake_case, unused_mut)]
+#[allow(
+    unused_braces,
+    unused_parens,
+    unused_variables,
+    non_snake_case,
+    unused_mut
+)]
 #[derive(Clone, Copy)]
 pub struct GrammarParser;
 unsafe impl ::std::marker::Send for GrammarParser {}
@@ -6754,14 +6756,14 @@ impl ::rusty_lr_core::parser::Parser for GrammarParser {
                         _ => unreachable!(),
                     };
                     rules
-                        .push(::rusty_lr_core::parser::state::RuleInfo {
+                        .push(::rusty_lr_core::parser::table::RuleInfo {
                             name,
                             len: RULE_LENGTHS[i] as usize,
                             precedence,
                         });
                 }
                 let num_states = 167usize;
-                let mut states = Vec::with_capacity(num_states);
+                let mut state_rows = Vec::with_capacity(num_states);
                 for i in 0..num_states {
                     let term_start = SHIFT_TERM_OFFSETS[i] as usize;
                     let term_end = SHIFT_TERM_OFFSETS[i + 1] as usize;
@@ -6835,10 +6837,10 @@ impl ::rusty_lr_core::parser::Parser for GrammarParser {
                         ruleset: Vec::new(),
                         can_accept_error,
                     };
-                    states.push(intermediate);
+                    state_rows.push(intermediate);
                 }
-                ::rusty_lr_core::parser::state::IntermediateTables {
-                    states,
+                ::rusty_lr_core::parser::table::IntermediateTables {
+                    state_rows,
                     rules,
                 }
                     .into()
@@ -6847,4 +6849,3 @@ impl ::rusty_lr_core::parser::Parser for GrammarParser {
 }
 
 // ==============================Generated Codes End===============================
-
