+++ scripts/diff/calculator.rs
@@ -37,9 +37,9 @@ pub type EContext = ::rusty_lr::parser::deterministic::Context<EParser, EDataSta
 /// type alias for CFG production rule
 #[allow(non_camel_case_types, dead_code)]
 pub type ERule = ::rusty_lr::rule::ProductionRule<ETerminalClasses, ENonTerminals>;
-/// type alias for DFA state
+/// type alias for runtime parser tables
 #[allow(non_camel_case_types, dead_code)]
-pub type EState = ::rusty_lr::parser::state::DenseState<
+pub type ETables = ::rusty_lr::parser::table::DenseFlatTables<
     ETerminalClasses,
     ENonTerminals,
     u8,
@@ -632,7 +632,9 @@ impl ::rusty_lr::parser::Parser for EParser {
     type Term = Token;
     type TermClass = ETerminalClasses;
     type NonTerm = ENonTerminals;
-    type State = EState;
+    type StateIndex = u8;
+    type ReduceRules = u8;
+    type Tables = ETables;
     const ERROR_USED: bool = false;
     fn precedence_types(level: u8) -> Option<::rusty_lr::rule::ReduceType> {
         #[allow(unreachable_patterns)]
@@ -641,71 +643,13 @@ impl ::rusty_lr::parser::Parser for EParser {
             _ => None,
         }
     }
-    fn get_rules() -> &'static [ERule] {
-        static RULES: std::sync::OnceLock<Vec<ERule>> = std::sync::OnceLock::new();
-        RULES
+    fn get_tables() -> &'static ETables {
+        static TABLES: std::sync::OnceLock<ETables> = std::sync::OnceLock::new();
+        TABLES
             .get_or_init(|| {
                 static RULE_NAMES: &[u32] = &[0, 0, 1, 1, 2, 2, 3, 4];
                 static RULE_PRECEDENCES: &[u32] = &[1, 0, 5, 0, 0, 0, 0, 0];
-                static RULE_TOKENS_DATA: &[u32] = &[
-                    1, 2, 1, 3, 3, 4, 3, 5, 0, 6, 7, 8, 1, 7, 14,
-                ];
-                static RULE_TOKENS_OFFSETS: &[u32] = &[0, 3, 4, 7, 8, 9, 12, 13, 15];
-                let num_rules = 8usize;
-                let mut rules = Vec::with_capacity(num_rules);
-                for i in 0..num_rules {
-                    let name = ENonTerminals::from_usize(RULE_NAMES[i] as usize);
-                    let prec_val = RULE_PRECEDENCES[i];
-                    let precedence = match prec_val & 3 {
-                        0 => None,
-                        1 => {
-                            Some(
-                                ::rusty_lr::rule::Precedence::Fixed(
-                                    (prec_val >> 2) as usize,
-                                ),
-                            )
-                        }
-                        2 => {
-                            Some(
-                                ::rusty_lr::rule::Precedence::Dynamic(
-                                    (prec_val >> 2) as usize,
-                                ),
-                            )
-                        }
-                        _ => unreachable!(),
-                    };
-                    let token_start = RULE_TOKENS_OFFSETS[i] as usize;
-                    let token_end = RULE_TOKENS_OFFSETS[i + 1] as usize;
-                    let mut rule = Vec::with_capacity(token_end - token_start);
-                    for idx in token_start..token_end {
-                        let val = RULE_TOKENS_DATA[idx];
-                        let is_nonterm = (val & 1) != 0;
-                        let sym_idx = (val >> 1) as usize;
-                        let token = if is_nonterm {
-                            ::rusty_lr::Token::NonTerm(
-                                ENonTerminals::from_usize(sym_idx),
-                            )
-                        } else {
-                            ::rusty_lr::Token::Term(
-                                ETerminalClasses::from_usize(sym_idx),
-                            )
-                        };
-                        rule.push(token);
-                    }
-                    rules
-                        .push(::rusty_lr::rule::ProductionRule {
-                            name,
-                            rule,
-                            precedence,
-                        });
-                }
-                rules
-            })
-    }
-    fn get_states() -> &'static [EState] {
-        static STATES: std::sync::OnceLock<Vec<EState>> = std::sync::OnceLock::new();
-        STATES
-            .get_or_init(|| {
+                static RULE_LENGTHS: &[u32] = &[3, 1, 3, 1, 1, 3, 1, 2];
                 static SHIFT_TERM_DATA: &[u32] = &[
                     2147516416, 65539, 2147516416, 65539, 2147614721, 2147516416, 65539,
                     229378, 2147516416, 65539, 327684, 2147876871,
@@ -732,8 +676,38 @@ impl ::rusty_lr::parser::Parser for EParser {
                 static CAN_ACCEPT_ERROR: &[u8] = &[
                     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 ];
+                let num_rules = 8usize;
+                let mut rules = Vec::with_capacity(num_rules);
+                for i in 0..num_rules {
+                    let name = ENonTerminals::from_usize(RULE_NAMES[i] as usize);
+                    let prec_val = RULE_PRECEDENCES[i];
+                    let precedence = match prec_val & 3 {
+                        0 => None,
+                        1 => {
+                            Some(
+                                ::rusty_lr::rule::Precedence::Fixed(
+                                    (prec_val >> 2) as usize,
+                                ),
+                            )
+                        }
+                        2 => {
+                            Some(
+                                ::rusty_lr::rule::Precedence::Dynamic(
+                                    (prec_val >> 2) as usize,
+                                ),
+                            )
+                        }
+                        _ => unreachable!(),
+                    };
+                    rules
+                        .push(::rusty_lr::parser::table::RuleInfo {
+                            name,
+                            len: RULE_LENGTHS[i] as usize,
+                            precedence,
+                        });
+                }
                 let num_states = 13usize;
-                let mut states = Vec::with_capacity(num_states);
+                let mut state_rows = Vec::with_capacity(num_states);
                 for i in 0..num_states {
                     let term_start = SHIFT_TERM_OFFSETS[i] as usize;
                     let term_end = SHIFT_TERM_OFFSETS[i + 1] as usize;
@@ -797,9 +771,13 @@ impl ::rusty_lr::parser::Parser for EParser {
                         ruleset: Vec::new(),
                         can_accept_error,
                     };
-                    states.push(intermediate.into());
+                    state_rows.push(intermediate);
                 }
-                states
+                ::rusty_lr::parser::table::IntermediateTables {
+                    state_rows,
+                    rules,
+                }
+                    .into()
             })
     }
 }
